+++ ChangeLog.md
@@ -1,4 +1,4 @@
-## 0.2.0.0 (unreleased)
+## 0.2.0.0
 
 * Clearer error messages on crash
 * Exception on invalid credentials